data-processing/tests/integration/test_database.py
"""
Integration tests for database operations
"""

import pytest
from src.config.settings import Settings
from src.config.database import DatabaseConfig

class TestDatabaseIntegration:
    """Test database integration"""
    
    @pytest.fixture
    def db_config(self):
        """Create database config"""
        settings = Settings()
        return DatabaseConfig(settings)
    
    def test_database_connection(self, db_config):
        """Test database connection"""
        # This would test actual database connection
        assert db_config is not None
    
    def test_create_tables(self, db_config):
        """Test table creation"""
        # This would test actual table creation
        db_config.create_tables()
